requirements.txt
azure-cognitiveservices-speech==1.33.0
fastapi==0.104.1
langchain==0.0.339
num2words==0.5.13
numpy==1.26.2
openai==1.3.4
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
starlette==0.27.0
uvicorn==0.24.0.post1
ffmpeg
